Move `FeedImageDataLoader` to its own file
